requirements.txt
pandas
python-telegram-bot
python-decouple
gspread
oauth2client